requirements.txt
# RunBot 核心依赖
pytest>=7.2.0
attrs>=23.0.0
requests>=2.28.0
pyyaml>=6.0
loguru>=0.6.0
tenacity>=8.0.0
jinja2>=3.1.0
click>=8.1.0
emoji>=2.2.0

# 报告
allure-pytest>=2.8.0
allure-python-commons>=2.8.0

# 数据提取与校验
jsonpath>=0.82
jsonschema>=4.17.0
genson>=1.2.2

# 数据库
psycopg2-binary>=2.9.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# AI 用例生成（可选）
openai>=1.0.0
